requirements.txt
alembic==1.8.0
asgiref==3.5.2
Authlib==0.15.5
Babel==2.10.3
backports.zoneinfo==0.2.1
bcrypt==3.2.2
bidict==0.22.0
blinker==1.4
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
click==8.1.3
cryptography==3.4.8
Django==4.0.5
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.2.1
eventlet==0.33.0
Flask==2.1.2
Flask-Babel==2.0.0
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.4
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
httpagentparser==1.9.3
idna==2.10
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
ldap3==2.9.1
Mako==1.2.0
MarkupSafe==2.1.1
paramiko==2.11.0
passlib==1.7.4
pgadmin4==6.10
Pillow==9.1.1
pkg_resources==0.0.0
psutil==5.9.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
pyotp==2.6.0
python-dateutil==2.8.2
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2021.3
qrcode==7.3.1
requests==2.25.1
s3transfer==0.5.2
simplejson==3.17.6
six==1.16.0
speaklater3==1.4
SQLAlchemy==1.4.39
sqlparse==0.4.2
sshtunnel==0.4.0
ua-parser==0.10.0
urllib3==1.26.9
user-agents==2.2.0
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.8.0
